pyproject.toml
[project]
name = "git-data-discord-bot"
version = "0.1.0"
description = "Discord bot for PR-merge notifications and team todos"
requires-python = ">=3.13"
dependencies = [
    "discord.py>=2.4",
    "aiohttp>=3.10",
    "aiosqlite>=0.20",
    "python-dotenv>=1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/bot"]

[tool.ruff]
line-length = 100
target-version = "py313"
